Extract collection building functionality.
